Adding `onerror` handler to `shutil.rmtree` function
To address a `PermissionError` on windows systems when invoking `shutil.rmtree` in `directory_cleaner`, the `onerror` function will change the permissions of the file to allow for write access. 

This may provide a brief buffer when deleting directories and files recursely due to a possible race condition; if the write permission error is actually present, then the `onerror` logic activates. Else, it serves as a passthrough.